LayoutTests/fast/canvas/webgl/context-attributes-expected.txt

Test passing down and fetching of WebGLContextAttributes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test default values
PASS context = create3DContext() is non-null.
PASS attribs = context.getContextAttributes() is non-null.
PASS attribs.depth is true
PASS attribs.alpha is true
PASS attribs.stencil == true || attribs.stencil == false is true
PASS attribs.antialias == true || attribs.antialias == false is true
PASS attribs.premultipliedAlpha is true
Test customized values
PASS context = create3DContext({ stencil: false, antialias: false }) is non-null.
PASS attribs = context.getContextAttributes() is non-null.
PASS attribs.depth is true
PASS attribs.alpha is true
PASS attribs.stencil is false
PASS attribs.antialias is false
PASS attribs.premultipliedAlpha is true
Test customized values
PASS context = create3DContext({ depth: false, stencil: true, antialias: false }) is non-null.
PASS attribs = context.getContextAttributes() is non-null.
PASS attribs.depth == attribs.stencil is true
PASS attribs.alpha is true
PASS attribs.antialias is false
PASS attribs.premultipliedAlpha is true
Test customized values
PASS context = create3DContext({ premultipliedAlpha: false, antialias: false }) is non-null.
PASS attribs = context.getContextAttributes() is non-null.
PASS attribs.depth is true
PASS attribs.alpha is true
PASS attribs.stencil == true || attribs.stencil == false is true
PASS attribs.antialias is false
PASS attribs.premultipliedAlpha is true
PASS successfullyParsed is true

TEST COMPLETE
